refactor(rating): rename functional component dir for consistency
