+++ packages/calcite-components/src/components/rating/functional/star.tsx
